requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.6
python-socketio==5.11.1
python-engineio==4.8.2
gunicorn==21.2.0
python-chess==1.999
psycopg2-binary
